feat: initialize database schema with all core system tables and migrations

+++ database/migrations/2026_05_29_030945_create_notifications_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->string('type', 50);
             $table->string('notifiable_type', 100);
-            $table->unsignedBigInteger('notifiable_id');
+            $table->unsignedInteger('notifiable_id');
             $table->index(['notifiable_type', 'notifiable_id']);
             $table->text('data');
             $table->timestamp('read_at')->nullable();
